Fix MACOS build

+++ fizzbuzz_server/timing.cc
@@ -8,7 +8,7 @@ TimeLog::TimeLog()
 	mEntries.reserve(MAX_TIME_LOG_SIZE);
 }
 
-uint64_t ToEpoch(Timestamp ts)
+long long ToEpoch(Timestamp ts)
 {
 	return std::chrono::duration_cast<std::chrono::nanoseconds>(ts.time_since_epoch()).count();
 }
@@ -20,7 +20,7 @@ void TimeLog::DumpLog()
 	fprintf(f, "Header: TimeLogEntry: request=uint64,startTS=timestamp,finishParsingTS=timestamp,finishProcessingTS=timestamp,finishSendTS=timestamp\n");
 	for (auto& entry : mEntries)
 	{
-		fprintf(f, "Data: TimeLogEntry: request=%lu,startTS=%lu,finishParsingTS=%lu,finishProcessingTS=%lu,finishSendTS=%lu\n",
+		fprintf(f, "Data: TimeLogEntry: request=%lld,startTS=%lld,finishParsingTS=%lld,finishProcessingTS=%lld,finishSendTS=%lld\n",
 			entry.request, ToEpoch(entry.startTS), ToEpoch(entry.finishParsingTS),
 			ToEpoch(entry.finishProcessingTS), ToEpoch(entry.finishSendTS));
 	}
